添加对战信息展示，增加宠物形态，扩展技能信息。

+++ PixePixie/ViewsAndControllers/MenuControllers/PPTableView.m
@@ -106,7 +106,7 @@
     if (self.choosePassNumber!=nil&&self.choosePassNumberSel!=nil&&[self.choosePassNumber respondsToSelector:self.choosePassNumberSel]) {
         
 
-        [self.choosePassNumber performSelectorInBackground:self.choosePassNumberSel withObject:[NSNumber numberWithInt:indexPath.row]];
+        [self.choosePassNumber performSelectorInBackground:self.choosePassNumberSel withObject:[NSNumber numberWithInt:(int)indexPath.row]];
     }
 //
 }
