+++ PixePixie/ViewsAndControllers/Views/PPPassNumberScroll.m
@@ -30,7 +30,7 @@
     
     
     UIScrollView *scrollContent=[[UIScrollView alloc] initWithFrame:CGRectMake(0.0f, 0.0f, self.frame.size.width, self.frame.size.height)];
-    NSLog(@"count=%d",passCount);
+    NSLog(@"count=%ld",(long)passCount);
     scrollContent.showsHorizontalScrollIndicator=NO;
     
     scrollContent.contentSize=CGSizeMake(self.frame.size.width*passCount, self.frame.size.height);
@@ -75,7 +75,7 @@
 -(void)passBtnClick:(UIButton *)sender
 {
     if (self.target!=nil&&self.selector!=nil&&[self.target respondsToSelector:self.selector]) {
-        [self.target performSelectorInBackground:self.selector withObject:[NSNumber numberWithInt:sender.tag]];
+        [self.target performSelectorInBackground:self.selector withObject:[NSNumber numberWithInt:(int)sender.tag]];
 
     }
     [self setHidden:YES];
